pyproject.toml
[project]
name = "tags-trending"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "orjson>=3.11.5",
    "polars>=1.36.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.5",
]
